rename ExceptionMappersRegistry to ExceptionMappersStorage
