+++ errors/src/commonMain/kotlin/dev/icerock/moko/errors/mappers/ConditionPair.kt
@@ -2,7 +2,7 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.errors.registry
+package dev.icerock.moko.errors.mappers
 
 data class ConditionPair(
     val condition: (BasicException) -> Boolean,
