+++ errors/src/commonMain/kotlin/dev/icerock/moko/errors/mappers/ExceptionMappersStorage.kt
@@ -2,7 +2,7 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.moko.errors.registry
+package dev.icerock.moko.errors.mappers
 
 import dev.icerock.moko.errors.MR
 import dev.icerock.moko.resources.desc.StringDesc
@@ -14,7 +14,7 @@ typealias BasicException = Throwable
 internal typealias ExceptionMapper = (BasicException) -> Any
 
 @ThreadLocal
-object ExceptionMappersRegistry {
+object ExceptionMappersStorage {
 
     var unknownErrorText: StringDesc = MR.strings.moko_errors_unknownError.desc()
 
@@ -27,7 +27,7 @@ object ExceptionMappersRegistry {
         resultClass: KClass<T>,
         exceptionClass: KClass<E>,
         mapper: (E) -> T
-    ): ExceptionMappersRegistry {
+    ): ExceptionMappersStorage {
         if (!mappersMap.containsKey(resultClass)) {
             mappersMap[resultClass] = mutableMapOf()
         }
@@ -38,7 +38,7 @@ object ExceptionMappersRegistry {
     fun <T : Any> register(
         resultClass: KClass<T>,
         conditionPair: ConditionPair
-    ): ExceptionMappersRegistry {
+    ): ExceptionMappersStorage {
         if (!conditionMappers.containsKey(resultClass)) {
             conditionMappers[resultClass] = mutableListOf()
         }
@@ -46,21 +46,35 @@ object ExceptionMappersRegistry {
         return this
     }
 
+    /**
+     * Register simple mapper (E) -> T.
+     */
     inline fun <reified E : BasicException, reified T : Any> register(
         noinline mapper: (E) -> T
-    ): ExceptionMappersRegistry {
-        return register(T::class, E::class, mapper)
+    ): ExceptionMappersStorage {
+        return register(
+            T::class,
+            E::class,
+            mapper
+        )
     }
 
+    /**
+     * Registers mapper (Throwable) -> T with specific condition (Throwable) -> Boolean.
+     */
     inline fun <reified T : Any> condition(
         noinline condition: (BasicException) -> Boolean,
         noinline mapper: (BasicException) -> T
-    ): ExceptionMappersRegistry = register(
-        resultClass = T::class,
-        conditionPair = ConditionPair(condition, mapper as ExceptionMapper)
-    )
+    ): ExceptionMappersStorage =
+        register(
+            resultClass = T::class,
+            conditionPair = ConditionPair(
+                condition,
+                mapper as ExceptionMapper
+            )
+        )
 
-    fun <T : Any, E : BasicException> find(
+    fun <E : BasicException, T : Any> find(
         resultClass: KClass<T>,
         exception: E,
         exceptionClass: KClass<out E>
@@ -71,15 +85,26 @@ object ExceptionMappersRegistry {
             ?: mappersMap.get(resultClass)?.get(exceptionClass) as? ((E) -> T)
     }
 
-    inline fun <reified T : Any, E : BasicException> find(exception: E): ((E) -> T)? =
-        find(T::class, exception, exception::class)
+    /**
+     * Tries to find mapper (E) -> T. First, a mapper with a condition is looked for. If this was
+     * not found, then a simple mapper is looked for.
+     */
+    inline fun <E : BasicException, reified T : Any> find(exception: E): ((E) -> T)? =
+        find(
+            T::class,
+            exception,
+            exception::class
+        )
 
-    fun setUnknownErrorText(text: StringDesc): ExceptionMappersRegistry {
+    fun setUnknownErrorText(text: StringDesc): ExceptionMappersStorage {
         unknownErrorText = text
         return this
     }
 }
 
-fun <E : BasicException> ExceptionMappersRegistry.throwableToStringDesc(e: E): StringDesc {
-    return find<StringDesc, E>(e)?.invoke(e) ?: unknownErrorText
+fun <E : BasicException> ExceptionMappersStorage.throwableToStringDesc(e: E): StringDesc {
+    return find<E, StringDesc>(
+        e
+    )
+        ?.invoke(e) ?: unknownErrorText
 }
