+++ sample/mpp-library/src/commonMain/kotlin/com/icerockdev/library/ExceptionStorage.kt
@@ -1,12 +1,12 @@
 package com.icerockdev.library
 
-import dev.icerock.moko.errors.registry.ExceptionMappersRegistry
+import dev.icerock.moko.errors.mappers.ExceptionMappersStorage
 import dev.icerock.moko.resources.desc.StringDesc
 import dev.icerock.moko.resources.desc.desc
 
 @Suppress("MagicNumber")
 fun initExceptionRegistry() {
-    ExceptionMappersRegistry
+    ExceptionMappersStorage
         .condition<StringDesc>(
             condition = { it is CustomException && it.code == 10 },
             mapper = { MR.strings.myExceptionText.desc() }
